Moved User to a module folder, moved BotUser class to the globals
