+++ lib/src/events/accordion_poll_events.dart
@@ -1,4 +1,4 @@
-import 'package:weather/src/core/user.dart' show BotUser;
+import 'package:weather/src/globals/bot_user.dart';
 
 class AccordionPollEvent {
   final BotUser fromUser;
