+++ lib/src/modules/accordion_poll/accordion_poll.dart
@@ -1,9 +1,9 @@
 import 'package:weather/src/globals/accordion_poll.dart';
 import 'package:weather/src/injector/injection.dart';
 import 'package:weather/src/core/event_bus.dart';
+import 'package:weather/src/globals/bot_user.dart';
 import 'package:weather/src/events/accordion_poll_events.dart';
 import 'package:weather/src/modules/chat/chat.dart';
-import 'package:weather/src/core/user.dart' show BotUser;
 
 class AccordionPoll {
   final Chat chat;
