+++ lib/src/modules/check_reminder_manager.dart
@@ -1,7 +1,7 @@
 import 'package:weather/src/core/check_reminder.dart';
-import 'package:weather/src/core/user.dart';
 import 'package:weather/src/globals/message_event.dart';
 import 'package:weather/src/injector/injection.dart';
+import 'package:weather/src/modules/user/user.dart';
 import 'package:weather/src/modules/chat/chat.dart';
 import 'package:weather/src/platform/platform.dart';
 import 'package:weather/src/utils/logger.dart';
