+++ lib/src/modules/user/user.dart
@@ -1,36 +1,8 @@
 import 'dart:async';
 import 'package:cron/cron.dart';
 import 'package:weather/src/injector/injection.dart';
-import 'repositories/bot_user_repository.dart';
-
-class BotUser {
-  final String id;
-  final bool isPremium;
-  final bool deleted;
-  final bool banned;
-  final bool moderator;
-
-  String name;
-
-  BotUser(
-      {required this.id,
-      required this.name,
-      required this.isPremium,
-      required this.deleted,
-      required this.banned,
-      required this.moderator}) {
-    var markedAsPremium = name.contains('⭐');
-
-    if (isPremium && !markedAsPremium) {
-      name += ' ⭐';
-    } else if (!isPremium && markedAsPremium) {
-      name = name.replaceAll(' ⭐', '');
-    }
-  }
-
-  Map<String, dynamic> toJson() =>
-      {'id': id, 'name': name, 'isPremium': isPremium, 'deleted': deleted, 'banned': banned, 'moderator': moderator};
-}
+import 'package:weather/src/core/repositories/bot_user_repository.dart';
+import 'package:weather/src/globals/bot_user.dart';
 
 class User {
   final BotUserRepository _userDb;
