+++ lib/src/platform/platform.dart
@@ -1,7 +1,5 @@
 import 'dart:async';
 
-import 'package:weather/src/core/user.dart';
-
 import 'package:weather/src/globals/chat_platform.dart';
 import 'package:weather/src/globals/message_event.dart';
 import 'package:weather/src/globals/bot_command.dart';
@@ -10,6 +8,7 @@ import 'package:weather/src/globals/accordion_poll.dart';
 import 'package:weather/src/platform/telegram_platform.dart';
 import 'package:weather/src/platform/discord_platform.dart';
 
+import 'package:weather/src/modules/user/user.dart';
 import 'package:weather/src/modules/chat/chat.dart';
 
 abstract class Platform<T> {
