+++ test/access_test.dart
@@ -3,7 +3,7 @@ import 'package:weather/src/core/access.dart';
 import 'package:weather/src/globals/access_level.dart';
 import 'package:weather/src/globals/chat_platform.dart';
 import 'package:weather/src/modules/chat/chat.dart';
-import 'package:weather/src/core/user.dart';
+import 'package:weather/src/modules/user/user.dart';
 import 'utils/setup.dart';
 import 'utils/helpers.dart';
