+++ test/user_test.dart
@@ -1,6 +1,6 @@
 import 'package:test/test.dart';
 import 'package:postgres/postgres.dart';
-import 'package:weather/src/core/user.dart';
+import 'package:weather/src/modules/user/user.dart';
 import 'package:weather/src/modules/chat/chat.dart';
 import 'package:weather/src/globals/chat_platform.dart';
 import 'utils/setup.dart';
